ECOM/requirements.txt
aerich==0.6.2
aiomysql==0.0.22
aiosqlite==0.17.0
alembic==1.7.6
amqp==5.0.9
aniso8601==7.0.0
anyio==3.5.0
appdirs==1.4.4
asgiref==3.5.0
attrs==21.4.0
automaton==2.4.0
autopage==0.5.0
Babel==2.9.1
bcrypt==3.2.0
beautifulsoup4==4.10.0
boto3==1.20.49
botocore==1.23.49
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
cliff==3.10.0
cmd2==2.3.3
cryptography==36.0.1
ddlparse==1.10.0
debtcollector==2.4.0
decorator==5.1.1
dictdiffer==0.9.0
Django==4.0.2
django-bootstrap3==21.2
djangorestframework==3.13.1
dnspython==2.2.0
docker==5.0.3
dogpile.cache==1.1.5
dynaconf==3.1.7
email-validator==1.1.3
eventlet==0.33.0
extras==1.0.0
fastapi==0.73.0
fastapi-login==1.7.3
fastapi-migrations==0.0.4
fasteners==0.17.3
fixtures==3.0.0
futurist==2.4.0
google-auth==2.6.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.1.2
h11==0.13.0
httptools==0.2.0
hvac==0.11.2
idna==3.3
install==1.3.5
iso8601==0.1.16
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
jsonpatch==1.32
jsonpointer==2.2
jsonschema==3.2.0
keystoneauth1==4.4.0
keystonemiddleware==9.3.0
knox==0.1.14
kombu==5.2.3
kubernetes==21.7.0
loguru==0.6.0
logutils==0.3.5
magnum==13.0.0
Mako==1.1.6
marathon==0.13.0
MarkupSafe==2.0.1
msgpack==1.0.3
munch==2.5.0
netaddr==0.8.0
netifaces==0.11.0
networkx==2.6.3
oauthlib==3.2.0
openstacksdk==0.61.0
orjson==3.6.6
os-client-config==2.1.0
os-service-types==1.7.0
osc-lib==2.4.2
oslo.cache==2.10.0
oslo.concurrency==4.5.0
oslo.config==8.7.1
oslo.context==4.0.0
oslo.db==11.1.0
oslo.i18n==5.1.0
oslo.log==4.6.1
oslo.messaging==12.12.0
oslo.metrics==0.4.0
oslo.middleware==4.5.0
oslo.policy==3.10.1
oslo.reports==2.3.0
oslo.serialization==4.2.0
oslo.service==2.8.0
oslo.upgradecheck==1.4.0
oslo.utils==4.12.1
oslo.versionedobjects==2.5.0
packaging==21.3
passlib==1.7.4
Paste==3.5.0
PasteDeploy==2.1.1
pbr==5.8.1
pecan==1.4.1
Pillow==9.0.1
prettytable==3.0.0
prometheus-client==0.13.1
promise==2.3
psutil==5.9.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycadf==3.1.1
pycparser==2.21
pydantic==1.9.0
pydot==1.4.2
pyinotify==0.9.6
PyJWT==2.3.0
PyMySQL==0.9.3
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyperclip==1.8.2
pypika-tortoise==0.1.3
pyrsistent==0.18.1
python-barbicanclient==5.2.0
python-cinderclient==8.2.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-glanceclient==3.5.0
python-heatclient==2.5.0
python-keystoneclient==4.4.0
python-multipart==0.0.5
python-neutronclient==7.7.0
python-novaclient==17.6.0
python-octaviaclient==2.4.0
python-openstackclient==5.7.0
python-slugify==5.0.2
python-swiftclient==3.13.0
pytz==2021.3
PyYAML==5.4.1
repoze.lru==0.7
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
requestsexceptions==1.4.0
rfc3986==2.0.0
Routes==2.5.1
rsa==4.8
Rx==1.6.1
s3transfer==0.5.1
shortuuid==1.0.8
simplegeneric==0.8.1
simplejson==3.17.6
singledispatch==3.7.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
SQLAlchemy==1.4.31
sqlalchemy-migrate==0.13.0
sqlparse==0.4.2
starlette==0.17.1
statsd==3.3.0
stevedore==3.5.0
taskflow==4.6.4
Tempita==0.5.2
tenacity==8.0.1
testresources==2.0.1
testscenarios==0.5.0
testtools==2.5.0
text-unidecode==1.3
tomlkit==0.8.0
tortoise-orm==0.18.1
typer==0.4.0
typing_extensions==4.0.1
ujson==5.1.0
urllib3==1.26.8
uvicorn==0.15.0
uvloop==0.16.0
validators==0.18.2
vine==5.0.0
waitress==2.0.0
warlock==1.3.3
watchgod==0.7
wcwidth==0.2.5
WebOb==1.8.7
websocket-client==1.2.3
websockets==10.1
WebTest==3.0.0
Werkzeug==2.0.3
wrapt==1.13.3
WSME==0.11.0
yappi==1.3.3
